requirements.txt
transformers
datasets
evaluate
huggingface_hub
python-dotenv
pandas
numpy